requirements.txt
Django==2.2.6
django-simple-history==2.7.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
